+++ Content.Shared/Body/Systems/SharedBodySystem.Organs.cs
@@ -6,20 +6,14 @@ using Content.Shared.Body.Organ;
 using Content.Shared.Body.Part;
 using Content.Shared.Body.Prototypes;
 using Content.Shared.Hands.EntitySystems;
-using Content.Shared.Interaction.Events;
-using Content.Shared.Medical;
-using Content.Shared.Mobs.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Prototypes;
-using System.Linq;
 
 namespace Content.Shared.Body.Systems;
 
 public partial class SharedBodySystem
 {
     [Dependency] private readonly SharedHandsSystem _hands = default!;
-    [Dependency] private readonly MobStateSystem _mobState = default!;
-    [Dependency] private readonly SharedRottingSystem _rotting = default!;
     [Dependency] private readonly SharedUserInterfaceSystem _ui = default!;
 
     private void InitializeOrgans()
@@ -28,11 +22,6 @@ public partial class SharedBodySystem
         SubscribeLocalEvent<OrganReplaceableComponent, OrganSelectionButtonPressedMessage>(OnOrganButton);
 
         SubscribeLocalEvent<OrganComponent, IsRottingEvent>(OnOrganIsRotting);
-
-        SubscribeLocalEvent<BodyComponent, DefibrillateAttemptEvent>(OnBodyDefib);
-        SubscribeLocalEvent<HeartComponent, UseInHandEvent>(OnHeartUseInHand);
-        SubscribeLocalEvent<HeartComponent, BodyOrganRelayedEvent<DefibrillateAttemptEvent>>(OnHeartDefib);
-        SubscribeLocalEvent<HeartComponent, StartedRottingEvent>(OnHeartRotting);
     }
 
     private void OnUIOpened(EntityUid uid, OrganReplaceableComponent component, BoundUIOpenedEvent args)
@@ -94,54 +83,6 @@ public partial class SharedBodySystem
     private void OnOrganIsRotting(EntityUid uid, OrganComponent component, ref IsRottingEvent args)
     {
         args.Handled = component.Body != null;
-    }
-
-    private void OnHeartUseInHand(EntityUid uid, HeartComponent component, UseInHandEvent args)
-    {
-        if (args.Handled)
-            return;
-
-        if (_rotting.IsRotten(uid))
-            return;
-
-        component.Beating ^= true;
-        _appearance.SetData(uid, HeartVisuals.Beating, component.Beating);
-        args.Handled = true;
-
-        if (component.Beating)
-            component.NextDamage = _timing.CurTime;
-    }
-
-
-    private void OnBodyDefib(EntityUid uid, BodyComponent component, DefibrillateAttemptEvent args)
-    {
-        if (!GetBodyOrganEntityComps<HeartComponent>((uid, component)).Any())
-        {
-            args.Cancel();
-            args.Reason = "defibrillator-heart-missing";
-            return;
-        }
-
-        RelayToOrgans(uid, component, args);
-    }
-
-    private void OnHeartDefib(EntityUid uid, HeartComponent component, ref BodyOrganRelayedEvent<DefibrillateAttemptEvent> args)
-    {
-        if (args.Args.Cancelled)
-            return;
-
-        if (component.Beating)
-            return;
-
-        args.Args.Cancel();
-        args.Args.Reason = "defibrillator-heart-off";
-    }
-
-    private void OnHeartRotting(EntityUid uid, HeartComponent component, StartedRottingEvent args)
-    {
-        _appearance.SetData(uid, HeartVisuals.Beating, false);
-        component.Beating = false;
-        component.NextDamage = _timing.CurTime;
     }
 
     private void AddOrgan(
